move tests out of orbits module
